Massive upgrade

+++ models/Team.py
@@ -5,7 +5,7 @@ class Team(db.Model):
     __tablename__ = 'teams'
 
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)  
-    fftId = db.Column(db.Integer, nullable=False)  
+    fftId = db.Column(db.BigInteger, nullable=False)
     player1Id = db.Column(db.Integer, db.ForeignKey('players.id'), nullable=False)  
     player2Id = db.Column(db.Integer, db.ForeignKey('players.id'), nullable=False)  
     rankingId = db.Column(db.Integer, db.ForeignKey('rankings.id'), nullable=False)  
@@ -16,6 +16,8 @@ class Team(db.Model):
     #relashionship
     player1 = db.relationship('Player', foreign_keys=[player1Id], backref='team_player1')
     player2 = db.relationship('Player', foreign_keys=[player2Id], backref='team_player2')
+    #matches_as_team1 = db.relationship('Match', foreign_keys=[Match.team1Id], backref='team1')
+    #matches_as_team2 = db.relationship('Match', foreign_keys=[Match.team2Id], backref='team2')
     ranking = db.relationship('Ranking', backref='teams')
 
     def __init__(self, fftId, player1Id, player2Id, rankingId, isActive):
